feat(standard-action-handler): ContextOptions에서 routeExternally 대신 openInlink, openOutlink를 받음
